python/react_agent/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.24",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
    "google-search-results>=2.4.2",
    "google-serp-api>=1.0.5",
]
